lib: Re-enable -Wall by default for lib.
This was disabled for lib to work with compilers that don't support it,
but not for Core and CGIs. Let's keep it on by default.

Silence the last -Wall warning in lib while we're at it.

+++ lib/worker.c
@@ -384,9 +384,9 @@ static void kill_job(child_process *cp, int reason)
 			 * reap attempt later.
 			 */
 			if (reason == ESTALE) {
-wlog("tv.tv_sec is currently %d", tv.tv_sec);
+				wlog("tv.tv_sec is currently %lu", (unsigned long)tv.tv_sec);
 				tv.tv_sec += 5;
-				wlog("Failed to reap child with pid %d. Next attempt @ %lu.%lu", cp->ei->pid, tv.tv_sec, tv.tv_usec);
+				wlog("Failed to reap child with pid %d. Next attempt @ %lu.%lu", cp->ei->pid, (unsigned long)tv.tv_sec, (unsigned long)tv.tv_usec);
 			} else {
 				tv.tv_usec = 250000;
 				if (tv.tv_usec > 1000000) {
